+++ lib/mc-markdown/formatters/image.rb
@@ -17,7 +17,7 @@ module MCMarkdown
         attrs    = match_data[2] || ""
 
         # check for attrs in class field
-        if attrs.include? ', '
+        if attrs.include? ':'
           attrs = attrs.split(', ').each_with_object([]) do |frag, out|
             frag = frag.split ':'
             out.push "#{frag[0].strip}='#{frag[1].strip}'"